requirements.txt
pytest
pytest-xdist
selenium
allure-pytest
webdriver-manager
urllib3
requests
pyparsing
packaging
iniconfig
idna
configparser
allure-python-commons
Faker
six
iniconfig
Appium-Python-Client
